- change all "unsigned {char,short,int,long}" to uint{8|16|32|32}_t - change all "u{char,short,int,long}" (from types.h) to uint{8|16|32|32}_t - remove types.h
git-svn-id: https://widelands.svn.sourceforge.net/svnroot/widelands/trunk@2761 37b2a8de-5219-0410-9f54-a31bc463ab9c

+++ src/cmd_queue.h
@@ -23,6 +23,7 @@
 #include "queue_cmd_ids.h"
 
 #include <queue>
+#include <stdint.h>
 
 class Editor_Game_Base;
 class Widelands_Map_Map_Object_Saver;
@@ -123,7 +124,7 @@ class Cmd_Queue {
 		 * independent of details of the priority_queue implementation.
 		 */
 		int category;
-		unsigned long serial;
+		uint32_t serial;
 
 		bool operator< (const cmditem& c) const
 		{
@@ -148,7 +149,7 @@ public:
 private:
 	Game *                       m_game;
 	std::priority_queue<cmditem> m_cmds;
-	unsigned long                nextserial;
+	uint32_t                     nextserial;
 };
 
 #endif // __S__CMD_QUEUE_H
